+++ example/configuration.tf
@@ -1,7 +1,0 @@
-provider "ironio" {
-  token = "${var.token}"
-}
-
-variable "token" {
-  description = "The IronMQ token"
-}