+++ example/data_source_projects.tf
@@ -1,0 +1,4 @@
+data "ironio_projects" "example" {
+  depends_on  = ["ironio_project.example"]
+  filter_name = "${ironio_project.example.name}"
+}