+++ example/input.tf
@@ -1,0 +1,3 @@
+variable "token" {
+  description = "The token"
+}