+++ example/provider_configuration.tf
@@ -1,0 +1,3 @@
+provider "ironio" {
+  token = "${var.token}"
+}