+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 FEATURES:
 
+* **New Data Source:** `ironio_projects`
 * **New Resource:** `ironio_project`
 * **New Resource:** `ironio_pull_queue`
 * **New Resource:** `ironio_push_queue`